pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel>=0.37"]
build-backend = "setuptools.build_meta"

[project]
name = "threadfactory"
version = "1.2.4"
description = "High-performance thread-safe (No-GIL–friendly) data structures and parallel operations for Python 3.13+."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.13"
license = "Apache-2.0"


authors = [
    { name = "Mark Geleta", email = "SynapticAISystems@gmail.com" }
]

maintainers = [
    { name = "Mark Geleta", email = "SynapticAISystems@gmail.com" }
]

keywords = [
    "concurrency",
    "parallelism",
    "thread-safe",
    "no-GIL",
    "threading",
    "parallel processing",
    "concurrent collections",
    "multithreading",
    "high-performance",
    "python concurrency",
    "free-threading",
    "thread factory",
    "ThreadFactory"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Distributed Computing",
    "Topic :: System :: Networking",
    "Topic :: Utilities",
    "Environment :: Console",
]

dependencies = ["python-ulid>=3.0.0"]

[project.urls]
Homepage = "https://github.com/Synaptic724/ThreadFactory"
Repository = "https://github.com/Synaptic724/ThreadFactory"
Issues = "https://github.com/Synaptic724/ThreadFactory/issues"
Documentation = "https://threadfactory.readthedocs.io/en/latest/"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["thread_factory*"]
exclude = [
    "tests*", "docs*", "examples*", "Plans*", "benchmarks*", "benchmark_data*",
    ".github*", ".venv*", "__pycache__", "*.egg-info", "TestIdeas"
]
